requirements.txt
# Core
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0

# ML
scikit-learn>=1.3.0
xgboost>=2.0.0
# pandas-ta optional (can use ta-lib or manual formulas)

# API clients
httpx>=0.25.0
feedparser>=6.0.0
aiohttp>=3.9.0
websockets>=12.0

# Polymarket
py-clob-client>=0.15.0

# Backtesting
# (custom backtester - no extra deps)

# Monitoring
prometheus-client>=0.19.0

# Utils
pydantic>=2.0.0
pyarrow>=14.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
